requirements.txt
# Core Libraries
numpy==2.0.2
scipy==1.14.1
pandas==2.2.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
tqdm==4.67.0

# Image Processing
pillow==11.0.0
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Transformers and NLP
transformers==4.46.3
tokenizers==0.20.3
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5

# Other Libraries
requests==2.32.3
safetensors==0.4.5
peft==0.13.2
supervision==0.25.0
pycocotools==2.0.8

# Plotting and Misc
wheel==0.45.1
pip==24.3.1
pexpect==4.9.0
mpmath==1.3.0
